requirements.txt
annotated-types==0.6.0
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.1
dnspython==2.6.1
exceptiongroup==1.2.1
iniconfig==2.0.0
packaging==24.0
pluggy==1.5.0
pydantic==2.7.1
pydantic_core==2.18.2
pymongo==4.7.2
pytest==8.2.0
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.11.0
